AbstractFuture#DEBUG_LEAKED_FUTURES adds additional details about the sources of leaked futures
Once I isolate the leak, this needs to be changed to `false`, since this currently slows down future constructors dramatically.

It appears that #setAsync is somehow failing to set a listener on the asyncFuture, nondeterministically. No idea how that's possible, given didSetListener is only set to `true` after we successfully do that, and if it fails, then we call asyncWork#end in a finally block :(