setListener now called after construtors are completed, instead of mid-constructor
This is error-prone, but also means that the futures are actually *fully constructed* before anything calls one of their methods.
This was the root of at least one bugs/leak. calling #catching on a completed future would call setListener, which would call #onFutureSucceeded, which could end up calling #toString, which would access members in the derived class which had not yet been initialized.

Related, I altered it so that AbstractFuture#toString no longer crashes when sourceClass() returns null.

Signed-off-by: MPD <mooing_psycho_duck@hotmail.com>